+++ hypervisor/release/npk_log.c
@@ -1,0 +1,10 @@
+/*
+ * Copyright (C) 2018 Intel Corporation. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#include <hypervisor.h>
+
+void npk_log_setup(__unused struct hv_npk_log_param *param) {}
+void npk_log_write(__unused const char *buf, __unused size_t len) {}
